update copy buttons, fix tests
