--- app/src/renderer/components/common/AnchorCopy.vue
@@ -1,22 +1,0 @@
-<template lang="pug">
-a.anchor-copy(:data-clipboard-text="value")
-  i.material-icons content_copy
-  .label {{ label }}
-</template>
-
-<script>
-import Clipboard from 'clipboard'
-export default {
-  name: 'anchor-copy',
-  mounted () {
-    let clipboard = new Clipboard(this.$el)
-    let self = this
-
-    clipboard.on('success', function (e) {
-      self.$store.commit('notify', {title: 'Copied to Clipboard', body: 'Selection has been successfully copied.'})
-      e.clearSelection()
-    })
-  },
-  props: ['value', 'label']
-}
-</script>
