+++ app/src/renderer/components/common/NiBtnCopy.vue
